Add new engines and rework HTEX for backward compatibility (#1057)
* No code changes beyond renaming and moving
* funcx_endpoint.executors.GCExecutorBase -> funcx_endpoint.engines.GlobusComputeEngineBase
* funcx_endpoint.executors.ProcessPoolExecutor -> funcx_endpoint.engines.ProcessPoolEngine
* funcx_endpoint.executors.GCExecutor -> funcx_endpoint.engines.GlobusComputeEngine
Similarly the helper script and init files have been moved and aggregated

* Updating tests to use the new engine names.

* HTEX no longer needs the funcx_service_address with changes from PR #1022

* Update the HTEX.start method to take additional params that are not available at init.

* * Update EPInterchange to require a single executor in config and use that
* `load_config` is removed, since that functionality
 is handled by `start_executor`
* `start_executor` now uses the updated `HTEX.start()` method.

* Stack of minor changes to tests that relied on older behavior.

* Moving over htex from executors to engines

* Mostly renaming
* funcx_endpoint.executors -> funcx_endpoint.engines
* HighThroughputExecutor -> HighThroughputEngine

* Add get_status_report to HTEX

* Updating tests following the executor->engine migration

* * EPInterchange to use queue.Queue for results_passthrough
* EPInterchange to require messages coming via results_passthrough to be messagepacked messages

* * Adding ThreadPoolEngine to module init
* Add ThreadPoolEngine to tests
* Adding a ThreadPoolEngine since the ProcessPoolEngine has issues on MacOS

* Updating HTEX engine to put mpacked Results in results_passthrough and consume mpacked Tasks via submit_raw

* * Adding a bunch of new tests
* Fixing older tests by pointing imports to funcx_endpoint.engines rather than funcx_endpoint.executors
* Some utility remote functions added to tests/utils.py
* Removing bad debug lines
* Minor bug fixes on exception reporting through results_passthrough

* * Minor bugfix. Pass along address kwarg to executor
* Set exception and start exiting thread in ReportingThread
* Aliasing `HighThroughputEngine` as `funcx_endpoint.executors.HighThroughputExecutor` for backward compatibility

* Updating parsl req to 2023.03.27 which has Ben's changes to the signal handler from parsl#2629

* * Ignoring wrong mypy error about address param being duplicated.
