+++ funcx_endpoint/funcx_endpoint/engines/__init__.py
@@ -1,4 +1,11 @@
 from funcx_endpoint.engines.globus_compute import GlobusComputeEngine
+from funcx_endpoint.engines.high_throughput.engine import HighThroughputEngine
 from funcx_endpoint.engines.process_pool import ProcessPoolEngine
+from funcx_endpoint.engines.thread_pool import ThreadPoolEngine
 
-__all__ = ["ProcessPoolEngine", "GlobusComputeEngine"]
+__all__ = [
+    "ProcessPoolEngine",
+    "GlobusComputeEngine",
+    "HighThroughputEngine",
+    "ThreadPoolEngine",
+]
