+++ funcx_endpoint/funcx_endpoint/engines/high_throughput/interchange.py
@@ -21,17 +21,17 @@ from funcx_common.tasks import ActorName, TaskState
 from parsl.version import VERSION as PARSL_VERSION
 
 from funcx.serialize import FuncXSerializer
-from funcx_endpoint.exception_handling import get_error_string, get_result_error_details
-from funcx_endpoint.executors.high_throughput.interchange_task_dispatch import (
+from funcx_endpoint.engines.high_throughput.interchange_task_dispatch import (
     naive_interchange_task_dispatch,
 )
-from funcx_endpoint.executors.high_throughput.messages import (
+from funcx_endpoint.engines.high_throughput.messages import (
     BadCommand,
     EPStatusReport,
     Heartbeat,
     Message,
     MessageType,
 )
+from funcx_endpoint.exception_handling import get_error_string, get_result_error_details
 from funcx_endpoint.logging_config import FXLogger
 
 if t.TYPE_CHECKING:
