+++ funcx_endpoint/funcx_endpoint/engines/high_throughput/zmq_pipes.py
@@ -6,7 +6,7 @@ import time
 import dill
 import zmq
 
-from funcx_endpoint.executors.high_throughput.messages import Message
+from funcx_endpoint.engines.high_throughput.messages import Message
 
 log = logging.getLogger(__name__)
