+++ funcx_endpoint/funcx_endpoint/engines/thread_pool.py
@@ -1,66 +1,63 @@
+from __future__ import annotations
+
 import logging
 import multiprocessing
-import os
 import typing as t
 import uuid
 from concurrent.futures import Future
+from concurrent.futures import ThreadPoolExecutor as NativeExecutor
+from multiprocessing.queues import Queue as mpQueue
 
+import psutil
 from funcx_common.messagepack.message_types import EPStatusReport, TaskTransition
-from parsl.executors.high_throughput.executor import HighThroughputExecutor
 
 from funcx_endpoint.engines.base import GlobusComputeEngineBase, ReportingThread
 
 logger = logging.getLogger(__name__)
 
 
-class GlobusComputeEngine(GlobusComputeEngineBase):
+class ThreadPoolEngine(GlobusComputeEngineBase):
     def __init__(
         self,
         *args,
-        label: str = "GlobusComputeEngine",
-        address: t.Optional[str] = None,
+        label: str = "ThreadPoolEngine",
         heartbeat_period: float = 30.0,
         **kwargs,
     ):
-        self.address = address
-        self.run_dir = os.getcwd()
         self.label = label
+        self.executor = NativeExecutor(*args, **kwargs)
         self._status_report_thread = ReportingThread(
             target=self.report_status, args=[], reporting_period=heartbeat_period
         )
         super().__init__(*args, heartbeat_period=heartbeat_period, **kwargs)
-        self.executor = HighThroughputExecutor(  # type: ignore
-            *args, address=address, **kwargs
-        )
 
     def start(
         self,
         *args,
         endpoint_id: t.Optional[uuid.UUID] = None,
-        run_dir: t.Optional[str] = None,
-        results_passthrough: t.Optional[multiprocessing.Queue] = None,
+        results_passthrough: t.Optional[mpQueue] = None,
         **kwargs,
-    ):
-        assert run_dir, "GCExecutor requires kwarg:run_dir at start"
-        assert endpoint_id, "GCExecutor requires kwarg:endpoint_id at start"
-        self.run_dir = os.path.join(os.getcwd(), run_dir)
-        self.endpoint_id = endpoint_id
-        self.executor.provider.script_dir = os.path.join(self.run_dir, "submit_scripts")
-        os.makedirs(self.executor.provider.script_dir, exist_ok=True)
-        if results_passthrough:
-            # Only update the default queue in GCExecutorBase if
-            # a queue is passed in
-            self.results_passthrough = results_passthrough
-        self.executor.start()
-        self._status_report_thread.start()
+    ) -> None:
+        """
 
-    def submit(
-        self,
-        func: t.Callable,
-        *args: t.Any,
-        **kwargs: t.Any,
-    ) -> Future:
-        return self.executor.submit(func, {}, *args, **kwargs)
+        Parameters
+        ----------
+        endpoint_id: Endpoint UUID
+        results_passthrough: Queue to which packed results will be posted
+        run_dir Not used
+
+        Returns
+        -------
+
+        """
+        assert endpoint_id, "ThreadPoolEngine requires kwarg:endpoint_id at start"
+        self.endpoint_id = endpoint_id
+        if results_passthrough:
+            self.results_passthrough = results_passthrough
+        assert self.results_passthrough
+
+        # mypy think the thread can be none
+        self._status_report_thread.start()
 
     def get_status_report(self) -> EPStatusReport:
         """
@@ -75,36 +72,52 @@ class GlobusComputeEngine(GlobusComputeEngineBase):
         executor_status: t.Dict[str, t.Any] = {
             "task_id": -2,
             "info": {
-                "total_cores": 0,
-                "total_mem": 0,
-                "new_core_hrs": 0,
+                "total_cores": multiprocessing.cpu_count(),
+                "total_mem": round(psutil.virtual_memory().available / (2**30), 1),
                 "total_core_hrs": 0,
-                "managers": 0,
-                "active_managers": 0,
-                "total_workers": 0,
-                "idle_workers": 0,
+                "total_workers": self.executor._max_workers,  # type: ignore
                 "pending_tasks": 0,
                 "outstanding_tasks": 0,
-                "worker_mode": 0,
-                "scheduler_mode": 0,
                 "scaling_enabled": False,
-                "mem_per_worker": 0,
-                "cores_per_worker": 0,
-                "prefetch_capacity": 0,
                 "max_blocks": 1,
                 "min_blocks": 1,
-                "max_workers_per_node": 0,
+                "max_workers_per_node": self.executor._max_workers,  # type: ignore
                 "nodes_per_block": 1,
                 "heartbeat_period": self._heartbeat_period,
             },
         }
         task_status_deltas: t.Dict[str, t.List[TaskTransition]] = {}
+
         return EPStatusReport(
             endpoint_id=self.endpoint_id,
             ep_status_report=executor_status,
             task_statuses=task_status_deltas,
         )
 
+    def submit(
+        self,
+        func: t.Callable,
+        *args: t.Any,
+        **kwargs: t.Any,
+    ) -> Future:
+        """We basically pass all params except the resource_specification
+        over to executor.submit
+        """
+        logger.warning("Got task")
+        return self.executor.submit(func, *args, **kwargs)
+
+    def status_polling_interval(self) -> int:
+        return 30
+
+    def scale_out(self, blocks: int) -> list[str]:
+        return []
+
+    def scale_in(self, blocks: int) -> list[str]:
+        return []
+
+    def status(self) -> dict:
+        return {}
+
     def shutdown(self):
         self._status_report_thread.stop()
-        return self.executor.shutdown()
+        self.executor.shutdown()
