+++ funcx_endpoint/tests/integration/funcx_endpoint/endpoint/test_interchange.py
@@ -40,8 +40,8 @@ def test_endpoint_id(funcx_dir):
         endpoint_id="mock_endpoint_id",
     )
 
-    for executor in ic.executors.values():
-        assert executor.endpoint_id == "mock_endpoint_id"
+    ic.start()
+    assert ic.executor.endpoint_id == "mock_endpoint_id"
 
 
 def test_start_requires_pre_registered(funcx_dir):
@@ -53,6 +53,7 @@ def test_start_requires_pre_registered(funcx_dir):
         )
 
 
+@pytest.mark.skip(reason="Obsolete. Executors are expected to return packed messages")
 def test_invalid_message_result_returned(mocker):
     ei = EndpointInterchange(
         config=Config(executors=[mocker.Mock(endpoint_id=None)]),
