+++ funcx_endpoint/tests/unit/test_endpointinterchange.py
@@ -21,9 +21,10 @@ def test_main_exception_always_quiesces(mocker, fs):
         return next(false_true_g)
 
     mocker.patch(f"{_mock_base}multiprocessing")
-    mocker.patch(f"{_mock_base}mpQueue")
+    mocker.patch(f"{_mock_base}queue")
+    mock_executor = MagicMock()
     ei = EndpointInterchange(
-        config=Config(executors=[]),
+        config=Config(executors=[mock_executor]),
         reg_info={"task_queue_info": {}, "result_queue_info": {}},
         reconnect_attempt_limit=num_iterations + 10,
     )
@@ -51,10 +52,11 @@ def test_reconnect_attempt_limit(mocker, fs, reconnect_attempt_limit):
         return next(false_true_g)
 
     mocker.patch(f"{_mock_base}multiprocessing")
-    mocker.patch(f"{_mock_base}mpQueue")
+    mocker.patch(f"{_mock_base}queue")
     mock_log = mocker.patch(f"{_mock_base}log")
+    mock_executor = MagicMock()
     ei = EndpointInterchange(
-        config=Config(executors=[]),
+        config=Config(executors=[mock_executor]),
         reg_info={"task_queue_info": {}, "result_queue_info": {}},
         reconnect_attempt_limit=reconnect_attempt_limit,
     )
