+++ funcx_endpoint/tests/unit/test_funcx_manager_unit.py
@@ -4,11 +4,11 @@ from unittest import mock
 
 from funcx_common.tasks import TaskState
 
-from funcx_endpoint.executors.high_throughput.funcx_manager import Manager as FXManager
-from funcx_endpoint.executors.high_throughput.messages import Task
+from funcx_endpoint.engines.high_throughput.funcx_manager import Manager as FXManager
+from funcx_endpoint.engines.high_throughput.messages import Task
 
 
-@mock.patch("funcx_endpoint.executors.high_throughput.funcx_manager.zmq")
+@mock.patch("funcx_endpoint.engines.high_throughput.funcx_manager.zmq")
 class TestFuncxManager:
     def test_task_to_worker_status_change(self, randomstring):
         task_type = randomstring()
