+++ funcx_endpoint/tests/unit/test_funcx_worker.py
@@ -5,8 +5,8 @@ from unittest import mock
 import pytest
 from funcx_common import messagepack
 
-from funcx_endpoint.executors.high_throughput.funcx_worker import FuncXWorker
-from funcx_endpoint.executors.high_throughput.messages import Task
+from funcx_endpoint.engines.high_throughput.funcx_worker import FuncXWorker
+from funcx_endpoint.engines.high_throughput.messages import Task
 
 
 def hello_world():
@@ -30,7 +30,7 @@ def ez_pack_function(serializer, func, args, kwargs):
 @pytest.fixture
 def test_worker():
     with mock.patch(
-        "funcx_endpoint.executors.high_throughput.funcx_worker.zmq.Context"
+        "funcx_endpoint.engines.high_throughput.funcx_worker.zmq.Context"
     ) as mock_context:
         # the worker will receive tasks and send messages on this mock socket
         mock_socket = mock.Mock()
